+++ services/serviceaccount/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.5.0
+- **Feature:** add support for Federated Identity Providers
+  - new operations: `CreateFederatedIdentityProvider`, `DeleteServiceFederatedIdentityProvider`, `ListFederatedIdentityProviders`,`PartialUpdateServiceAccountFederatedIdentityProvider`
+  - new models: `CreateFederatedIdentityProviderPayload`, `CreateFederatedIdentityProviderPayloadAssertionsInner`, `CreateFederatedIdentityProviderResponse`, `CreateFederatedIdentityProviderResponseAssertionsInner`, `FederatedListFederatedIdentityProvidersResponse`, `PartialUpdateServiceAccountFederatedIdentityProviderPayload`
+
 ## v0.4.2
 - **Bugfix:** Prevent year 0 timestamp issue